Fixed issue where a template would not have a ``_v_macros`` attribute.
This volatile attribute is relied upon by some legacy code such as the
Zope 2 product ``PageTemplates``.

In this changeset, the interface is changed to match that of the TAL
parser's ``getCode`` method which returns a tuple ``(program,
macros)``. These are in turn assigned to the volatile attributes
``_v_program`` and ``_v_macros``.