Workaround for CPU-burning pt_errors() on a recursive template.
See https://bugs.launchpad.net/zope.pagetemplate/+bug/732972

This implements the short workaround version of the solution mentioned in the
bug: let's skip macro expansion checking while we're formatting an error
traceback (with the default recursion limit of 100 this still burns a lot of
CPU: ~20 seconds on a 2.5 GHz Core i5).

The test can be easily extended for the better suggested solution (some time in
the future): just drop check_macro_expansion=False and make sure METAL silently
stops recursing iff it notices a loop and TAL evaluation is disabled.